\name{dynamicGraph-package}
\alias{dynamicGraph-package}
%\alias{dynamicGraph}
\docType{package}
\title{dynamicGraph}
\description{
  The package provides an interactive graphical tool
  for manipulating graphs by a Tcl/tk interface.
}
\details{
   \tabular{ll}{
      Package: \tab dynamicGraph \cr
      Type:    \tab Package      \cr
      Version: \tab 0.2.2.5      \cr
      Date:    \tab 2009-04-03   \cr
      License: \tab GPL (>= 2)   \cr
   }
}
\author{
Jens Henrik Badsberg

Maintainer: Jens Henrik Badsberg <coco@badsberg.eu>
}
% \references{ }
\keyword{ package }
